Mapping for full responses and filtering shortforms
